refactor: Update saveChecklistSchema for checklist validation
